Moved common config value paths to common
