+++ monkey/monkey_island/cc/services/attack/technique_reports/T1065.py
@@ -4,7 +4,7 @@ from monkey_island.cc.services.config import ConfigService
 
 __author__ = "VakarisZ"
 
-from monkey_island.cc.services.config_schema.config_value_paths import CURRENT_SERVER_PATH
+from common.config_value_paths import CURRENT_SERVER_PATH
 
 
 class T1065(AttackTechnique):
