+++ monkey/monkey_island/cc/services/configuration/utils.py
@@ -1,5 +1,5 @@
 from monkey_island.cc.services.config import ConfigService
-from monkey_island.cc.services.config_schema.config_value_paths import INACCESSIBLE_SUBNETS_PATH
+from common.config_value_paths import INACCESSIBLE_SUBNETS_PATH
 
 
 def get_config_network_segments_as_subnet_groups():
