+++ monkey/monkey_island/cc/services/zero_trust/scoutsuite/test_scoutsuite_auth_service.py
@@ -6,7 +6,7 @@ import dpath.util
 from monkey_island.cc.database import mongo
 from monkey_island.cc.server_utils import encryptor
 from monkey_island.cc.services.config import ConfigService
-from monkey_island.cc.services.config_schema.config_value_paths import AWS_KEYS_PATH
+from common.config_value_paths import AWS_KEYS_PATH
 from monkey_island.cc.services.zero_trust.scoutsuite.scoutsuite_auth_service import is_aws_keys_setup
 from monkey_island.cc.test_common.fixtures import FixtureEnum
